+++ HISTORY.md
@@ -2,7 +2,7 @@
 
 ## 0.1.8 (2019-05-05)
 
-* :issue:`8` Include capability to optimize alpha parameter
+* [#8 Include capability to optimize alpha parameter](https://github.com/bellockk/alphashape/issues/8)
 
 ## 0.1.7 (2019-04-26)
 
@@ -10,11 +10,11 @@
 
 ## 0.1.6 (2019-04-24)
 
-* :issue:`6` Include Jupyter Notebook in Examples
+* [#6 Include Jupyter Notebook in Examples](https://github.com/bellockk/alphashape/issues/6)
 
 ## 0.1.5 (2019-04-24)
 
-* :issue:`5` Create an Example Gallery in the Documentation
+* [#5 Create an Example Gallery in the Documentation](https://github.com/bellockk/alphashape/issues/5)
 
 ## 0.1.4 (2019-04-24)